Added some more comments to the code.

+++ src/VagrantTransient.php
@@ -1,17 +1,16 @@
 <?php
- /**
- *  
+/**
  * The MIT License (MIT)
- * 
+ *
  * Copyright (c) 2015  Philip Elson <phil@pegasus-commerce.com>
- * 
+ *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- * 
+ *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
 
@@ -26,6 +25,13 @@
  * Date: 23/06/15
  * Time: 11:25
  *
+ * PHP version 5.3+
+ *
+ * @category Pegasus_Utilities
+ * @package  VagrantTransient
+ * @author   Philip Elson <phil@pegasus-commerce.com>
+ * @license  MIT http://opensource.org/licenses/MIT
+ * @link     http://pegasus-commerce.com
  */
 namespace Pegasus;
 
@@ -42,10 +48,19 @@ use Symfony\Component\Console\Helper\ProgressBar;
 use Monolog\Logger;
 use Monolog\Handler\StreamHandler;
 
+/**
+ * This class provides the core methods for VagrantTransient
+ *
+ * @category Pegasus_Utilities
+ * @package  VagrantTransient
+ * @author   Philip Elson <phil@pegasus-commerce.com>
+ * @license  MIT http://opensource.org/licenses/MIT
+ * @link     http://pegasus-commerce.com
+ */
 class VagrantTransient extends Command
 {
     /**
-     * This is the default file where the environments are stored.
+     * This is the default file where the _environments are stored.
      */
     const DEFAULT_STORE_NAME    = "/.VagrantTransient";
 
@@ -59,14 +74,14 @@ class VagrantTransient extends Command
      *
      * @var null
      */
-    private $storeName          = null;
+    private $_storeName          = null;
 
     /**
-     * Array which holds the environments
+     * Array which holds the _environments
      *
      * @var array
      */
-    private $environments       = array();
+    private $_environments       = array();
 
     /**
      * Terminal output
@@ -82,7 +97,14 @@ class VagrantTransient extends Command
     protected $pwd              = null;
 
     /**
+     * String which represents the current environment
+     */
+    protected $currentEnvironment = null;
+
+    /**
      * Configures the application
+     *
+     * @return void
      */
     protected function configure()
     {
@@ -91,7 +113,7 @@ class VagrantTransient extends Command
             ->setDescription($this->getDescription())
             ->addOption(
                 'storage',
-                 null,
+                null,
                 InputOption::VALUE_OPTIONAL,
                 'Environment Location Storage',
                 $this->getDefaultFileName()
@@ -122,22 +144,31 @@ class VagrantTransient extends Command
 
     /**
      * This method is called before runTransient()
+     *
+     * @return $this
      */
     public function before()
     {
+        $this->setCurrentEnvironment($this->getDefaultCurrentEnvironment());
         $this->loadEnvironments();
+        return $this;
     }
 
     /**
      * Method runs the instance specific code.
+     *
+     * @return void
+     * @throws \Exception when not implemented
      */
     public function runTransient()
     {
-        throw new Exception('Not Yet Implemented');
+        throw new \Exception('Not Yet Implemented');
     }
 
     /**
      * This is the method which is called after the runTransient.
+     *
+     * @return void
      */
     public function after()
     {
@@ -147,14 +178,16 @@ class VagrantTransient extends Command
     /**
      * This is the command which executes the application
      *
-     * @param InputInterface $input
-     * @param OutputInterface $output
+     * @param InputInterface  $input  input interface for  command
+     * @param OutputInterface $output output interface for command
+     * 
+     * @return void
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         $this->pwd = getcwd();
         $this->output = $output;
-        $this->storeName = $input->getOption('storage');
+        $this->_storeName = $input->getOption('storage');
         $this->loadOutputStyles();
         $this->before();
         $this->runTransient();
@@ -163,55 +196,57 @@ class VagrantTransient extends Command
 
     /**
      * This method loads the command styles (wanring|general|notice|fatal_error)
+     *
+     * @return void
      */
     public function loadOutputStyles()
     {
-        $style = new OutputFormatterStyle('white', 'red', array('bold'));
+        $styleOne = array('bold');
+        $styleTwo = array('bold', 'underscore');
+        $style = new OutputFormatterStyle('white', 'red', $styleOne);
         $this->output->getFormatter()->setStyle('warning', $style);
-        $style = new OutputFormatterStyle('white', 'blue', array('bold'));
+        $style = new OutputFormatterStyle('white', 'blue', $styleOne);
         $this->output->getFormatter()->setStyle('general', $style);
-        $style = new OutputFormatterStyle('white', 'green', array('bold'));
+        $style = new OutputFormatterStyle('white', 'green', $styleOne);
         $this->output->getFormatter()->setStyle('notice', $style);
-        $style = new OutputFormatterStyle('white', 'red', array('bold', 'underscore'));
+        $style = new OutputFormatterStyle('white', 'red', $styleTwo);
         $this->output->getFormatter()->setStyle('fatal_error', $style);
     }
 
     /**
-     * This method loads the environments from disk into an array
+     * This method loads the _environments from disk into an array
      *
      * @return array
      */
     protected function loadEnvironments()
     {
-        $this->environments = array();
-        if(false == file_exists($this->getFileName()))
-        {
+        $this->_environments = array();
+        if (false == file_exists($this->getFileName())) {
             touch($this->getFileName());
         }
         $tempEnvironments = file($this->getFileName());
-        foreach($tempEnvironments as $key => $value)
-        {
+        foreach ($tempEnvironments as $key => $value) {
             $value = trim($value);
-            if(null != $value && '' != $value && false == is_array($value))
-            {
-                $this->environments[] = $value;
+            if (null != $value && '' != $value && false == is_array($value)) {
+                $this->_environments[] = $value;
             }
         }
-        return $this->environments;
+        return $this->_environments;
     }
 
     /**
-     * This method iterates over the environments writing them to disk.
+     * This method iterates over the _environments writing them to disk.
      *
      * @return string
      */
     public function save()
     {
-        unlink($this->getFileName());
+        if (true == file_exists($this->getFileName())) {
+            unlink($this->getFileName());
+        }
         touch($this->getFileName());
         $content = "";
-        foreach($this->getEnvironments() as $environment)
-        {
+        foreach ($this->getEnvironments() as $environment) {
             $content .= "{$environment}\n";
         }
         file_put_contents($this->getFileName(), $content);
@@ -219,7 +254,9 @@ class VagrantTransient extends Command
     }
 
     /**
-     * This method reloads the environments
+     * This method reloads the _environments
+     *
+     * @return void
      */
     public function refresh()
     {
@@ -244,18 +281,21 @@ class VagrantTransient extends Command
      */
     public function getFileName()
     {
-        if(null == $this->storeName) {
-            $this->storeName = $this->getDefaultFileName();
+        if (null == $this->_storeName) {
+            $this->_storeName = $this->getDefaultFileName();
         }
-        return $this->storeName;
+        return $this->_storeName;
     }
 
     /**
-     * This method purges the environments array
+     * This method purges the _environments array
+     *
+     * @return environments
      */
     public function purge()
     {
-        $this->environments = array();
+        $this->_environments = array();
+        return $this->environments;
     }
 
     /**
@@ -265,38 +305,38 @@ class VagrantTransient extends Command
      */
     public function getEnvironments()
     {
-        return $this->environments;
+        return $this->_environments;
     }
 
     /**
-     * This method overrides the environments array.
+     * This method overrides the _environments array.
      *
-     * @param array $environments
+     * @param array $_environments Is the new array of environments
+     *
      * @return $this
      */
-    public function setEnvironments(array $environments)
+    public function setEnvironments(array $_environments)
     {
-        $this->environments = $environments;
+        $this->_environments = $_environments;
         return $this;
     }
 
     /**
      * This method adds the environment to storage if it is not already in storage.
      *
-     * @param $name
+     * @param string $name is the name of the environment to create.
+     *
      * @return $this
      */
     public function createEnvironment($name)
     {
         $environments = $this->getEnvironments();
         $changed = false;
-        if(false == $this->environmentExists($name))
-        {
+        if (false == $this->environmentExists($name)) {
             $environments[] = $name;
             $changed = true;
         }
-        if(true == $changed)
-        {
+        if (true == $changed) {
             $this->setEnvironments($environments);
             $environments = array_reverse($environments);
             $this->setEnvironments($environments);
@@ -307,7 +347,8 @@ class VagrantTransient extends Command
     /**
      * This method returns true if the environment exists.
      *
-     * @param $name
+     * @param string $name Is the name of the environment to be checked
+     *
      * @return bool
      */
     public function environmentExists($name)
@@ -317,18 +358,17 @@ class VagrantTransient extends Command
 
     /**
      * This method removed an environment from the system.
-     * The environment must match exectly to the path in storage.
+     * The environment must match exactly to the path in storage.
      *
-     * @param $name
+     * @param string $name Is the name of the environment to be removed
+     *
      * @return $this
      */
     public function removeEnvironment($name)
     {
         $environments = $this->getEnvironments();
-        if(true == in_array($name, $environments))
-        {
-            if(($key = array_search($name, $environments)) !== false)
-            {
+        if (true == in_array($name, $environments)) {
+            if (($key = array_search($name, $environments)) !== false) {
                 unset($environments[$key]);
             }
         }
@@ -344,24 +384,20 @@ class VagrantTransient extends Command
      */
     protected function environmentsDown()
     {
-        foreach($this->getEnvironments() as $environment)
-        {
-            if(true == file_exists($environment))
-            {
+        foreach ($this->getEnvironments() as $environment) {
+            if (true == file_exists($environment)) {
                 //echo $environment;
                 chdir($environment);
                 $this->printLn("Shutting down vagrant in {$environment}");
                 $output = '';
                 $return = '';
                 exec('vagrant halt', $output, $return);
-                if(false == is_array($output))
-                {
+                if (false == is_array($output)) {
                     $output = array($output);
                 }
-                $this->printLn(implode(', ', $output), 0 == $return ? 'notice' : 'warning');
-            }
-            else
-            {
+                $style = (0 == $return) ? 'notice' : 'warning';
+                $this->printLn(implode(', ', $output), $style);
+            } else {
                 $this->printLn("Dir not found {$environment}");
 
             }
@@ -373,13 +409,14 @@ class VagrantTransient extends Command
     /**
      * This method prints a line to the terminal.
      *
-     * @param $message      Is the message
-     * @param string $type  Is the type of message (warning|notice|general)
+     * @param string $message Is the message
+     * @param string $type    Is the type of message (warning|notice|general)
+     *
+     * @return void
      */
     protected function printLn($message, $type='general')
     {
-        if(null != $this->output)
-        {
+        if (null != $this->output) {
             $message = (null == $type) ? $message : "<{$type}>{$message}</{$type}>";
             $this->output->writeLn($message);
         }
@@ -388,26 +425,49 @@ class VagrantTransient extends Command
     /**
      * Returns the directory of this environment.
      *
-     * @return string
+     * @return string which represents the current environment
      */
-    protected function getCurrentEnvironment()
+    public function getCurrentEnvironment()
+    {
+        return $this->currentEnvironment;
+    }
+
+    /**
+     * This method sets the current environment
+     *
+     * @param string $currentEnvironment is the environment name to use
+     * as the default.
+     *
+     * @return $this
+     */
+    public function setCurrentEnvironment($currentEnvironment)
+    {
+        $this->currentEnvironment = $currentEnvironment;
+        return $this;
+    }
+
+    /**
+     * This method returns the default current environment.
+     *
+     * @return string which represents the default environment
+     */
+    public function getDefaultCurrentEnvironment()
     {
         return __DIR__;
     }
 
     /**
-     * This method iterates over the environments. If there is no VagrantFile within that environment
-     * then it is removed from the list of environments.
+     * This method iterates over the _environments. If there is no VagrantFile
+     * within that environment then it is removed from the list of _environments.
      *
      * @return $this
      */
     protected function cleanEnvironments()
     {
         $environmentsWhichExist = array();
-        foreach($this->getEnvironments() as $key => $environment)
-        {
-            if(true == file_exists($environment.DIRECTORY_SEPARATOR.self::VAGRANT_FILE_NAME))
-            {
+        foreach ($this->getEnvironments() as $key => $environment) {
+            $fileName = $environment.DIRECTORY_SEPARATOR.self::VAGRANT_FILE_NAME;
+            if (true == file_exists($fileName)) {
                 $environmentsWhichExist[] = $environment;
             }
         }
